// src/medical-network/interfaces/data-csv.sheet.interface.ts
export interface IDataCsv {
  tier: string;
  tierRank: string;
  city: string;
  area: string;
  providerType: string;
  provider: string;
  speciality: string;
  subSpeciality: string;
  address: string;
  email: string;
  fax: string;
  phoneNumber: string;
  hasChronicMedication: string;
  websiteUrl: string;
  longitude: string;
  latitude: string;
  isOnline: string;
}